functional tests: disable TestVolumeSysfs on kvm
This test fails because the mounts are done on the host so they can't
be detected inside the VM.

Since mounting `/sys` doesn't really make sense on a kvm container,
let's disable this test.
